Add support for Invites behind HTTP Basic Auth, on Linux
Sometimes, invite URLs might be held on a private server, requiring
a username and password. This patch set implements support for these
URLs on Linux, by using curl to see whether the target URL is password
protected or not, and using Zenity to prompt for a password to pass to
Curl (via a temporary netrc file).

Note that this implementation does not support usernames or passwords
with a space in them, due to deficiencies in the netrc specification.